1. cparser make enum lookup upcast to s64 2. m64 type fix for unsigned s32 bits 3. backend adding settings parser use refleciton 4. cleaning up vt files and function before implementing vt evict system

+++ src/cparser/main.h
@@ -986,7 +986,7 @@ void WriteMetaFile(const s8* src_file_string,const s8* header_file_string,Generi
         
         u64 MetaGetEnumValueByName(MetaEnumEntry* array,u32 count,const s8* name);
         
-        void MetaGetEnumNamesByValue(u32 value,MetaEnumEntry* array,u32 count,const s8** names_array,u32* names_count);
+        void MetaGetEnumNamesByValue(u64 value,MetaEnumEntry* array,u32 count,const s8** names_array,u32* names_count);
         
         MetaDataStructEntry* MetaGetStructByNameHash(u32 hash);
         
@@ -1081,7 +1081,7 @@ void WriteMetaFile(const s8* src_file_string,const s8* header_file_string,Generi
     return MetaGetEnumValueByNameHash(array,count,PHashString(name));
     }
     
-    void MetaGetEnumNamesByValue(u32 value,MetaEnumEntry* array,u32 count,const s8** names_array,u32* names_count){
+    void MetaGetEnumNamesByValue(u64 value,MetaEnumEntry* array,u32 count,const s8** names_array,u32* names_count){
     
     u32 c = 0;
